Move theme and hooks to common
