+++ frontend/src/features/tasks/components/TaskTile.tsx
@@ -1,10 +1,10 @@
+import { useIsMobile } from "@/common/hooks/useIsMobile";
 import { Tile } from "@/components/composites/Tile";
 import { Tabs, TabsContent, TabsList, TabsTrigger } from "@/components/ui/tabs";
 import { TaskAnalyticsCard } from "@/features/tasks/components/TaskAnalyticsCard";
 import { TaskApprovalsCard } from "@/features/tasks/components/TaskApprovalsCard";
 import { TaskCommentsCard } from "@/features/tasks/components/TaskCommentsCard";
 import { TaskDetailsCard } from "@/features/tasks/components/TaskDetailsCard";
-import { useIsMobile } from "@/hooks/use-mobile";
 import { Task } from "@/types/task";
 import { EllipsisIcon, Share2Icon, UserPlus2Icon } from "lucide-react";
